+++ app/lib/json_web_token.rb
@@ -1,6 +1,6 @@
 class JsonWebToken
   class << self
-    ENCRYPTION_KEY = Rails.application.secrets[:secret_key_base]
+    ENCRYPTION_KEY = Rails.application.credentials.dig(:secret_key_base)
 
     def encode(payload, exp = CONFIG.dig(:authentication, :customer, :jwt_expiration_hours))
       payload[:exp] = exp.to_i.hours.from_now.to_i