+++ app/mailers/devise_sendgrid_mailer.rb
@@ -5,7 +5,7 @@ class DeviseSendgridMailer < Devise::Mailer
 
   def confirmation_instructions(record, token, opts={})
     if record.instance_of? User
-      link_url = api_ui_v1_user_confirmation_path(confirmation_token: token)
+      link_url = api_ui_v1_user_confirmation_url(confirmation_token: token)
       SendgridMailer.account_confirmation(record, link_url: link_url).deliver
     else
       super
@@ -14,7 +14,7 @@ class DeviseSendgridMailer < Devise::Mailer
 
   def reset_password_instructions(record, token, opts={})
     if record.instance_of? User
-      link_url = edit_api_ui_v1_user_password_path(reset_password_token: token)
+      link_url = edit_api_ui_v1_user_password_url(reset_password_token: token)
       SendgridMailer.reset_password(record, link_url: link_url).deliver
     else
       super